- Assets: YearRank est un decimal, ce qui simplifie les périodicités mensuelles, trimestrielles, etc. - Assets: EditorPageMethod: Nouveau bouton "Simulation" & Co, fonctionnel.
git-svn-id: https://svn.opac.ch/svn/cr/branches/vs2010.1@24022 787ff09c-d2b5-7842-b444-8dd0a418631d

+++ Epsitec.Cresus/Products/Assets.Server/BusinessLogic/AmortizationDefinition.cs
@@ -9,7 +9,7 @@ namespace Epsitec.Cresus.Assets.Server.BusinessLogic
 {
 	public struct AmortizationDefinition
 	{
-		public AmortizationDefinition(Guid expressionGuid, decimal rate, int yearRank, decimal yearCount, Periodicity periodicity, ProrataType prorataType, decimal round, decimal residual)
+		public AmortizationDefinition(Guid expressionGuid, decimal rate, decimal yearRank, decimal yearCount, Periodicity periodicity, ProrataType prorataType, decimal round, decimal residual)
 		{
 			this.ExpressionGuid  = expressionGuid;
 			this.Rate            = rate;
@@ -75,11 +75,11 @@ namespace Epsitec.Cresus.Assets.Server.BusinessLogic
 		}
 
 
-		public static AmortizationDefinition Empty = new AmortizationDefinition (Guid.Empty, 0.0m, 0, 0.0m, 0, 0.0m, 0.0m, 0.0m);
+		public static AmortizationDefinition Empty = new AmortizationDefinition (Guid.Empty, 0.0m, 0.0m, 0.0m, 0, 0.0m, 0.0m, 0.0m);
 
 		public readonly Guid					ExpressionGuid;
 		public readonly decimal					Rate;
-		public readonly int						YearRank;
+		public readonly decimal					YearRank;
 		public readonly decimal					YearCount;
 		public readonly Periodicity				Periodicity;
 		public readonly ProrataType				ProrataType;
